+++ CHANGELOG.md
@@ -28,6 +28,7 @@ before updating.
 - Fix various bugs related to networks and chunk loading/unloading.
 - Fix not being able to move network devices with mods like Carrier.
 - Fix item quantity not being formatted in the Grid.
+- Fix amount in detailed Grid tooltip being formatted with units.
 
 ### Changed